Set up Projects for Domain, API, and Entity Framework
Set up separate projects for the domain layer, MyAnimeList Api services, and entity framework. I also added HTTPClientFactory to dependency injection and used it to create an HTTP client to access the MyAnimeList Api. I get the data back in a json string, but I'm having trouble deserializing it. For some reason when it deserializes it makes all the properties of the anime null.