Created anime details page
Made it so that clicking on an anime card on the home page takes you to the anime details page with that anime's details. I still need to style it and add the rest of the details. I also need to figure out the most efficient method of getting the anime data. Right now I'm calling the myanimelist api every time someone visits the page, but that is a lot of api calls. I will probably eventually store the anime data in my own entity framework database and have it automatically update my database to synchronize my database with the myanimelist database. This way, I would drastically reduce the number of calls to the myanimelist api.