Created Search Functionality
Created search functionality using the search bar in the nav bar. The user can now search for anime in the search bar which will take them to a search results page. They can click on a search result which will take them to the anime details page for that anime.