Stored user info in session variables after login
In the authenticator, after logging a user in or creating an account, the users info is put into session variables are accessible as long as the session is lasts. I still need to figure out how to persist the information between sessions (every time you close the page and come back you have to log in again).