Started setting up authentication persistence
Began setting up authentication session persistence, but ran into some issues. I wanted to serialize the AuthCredential into a cookie and deserialize it, then use it to login via the FirebaseAuthClient, but you can't deserialize the AuthCredential class. I need to a way to authenticate the user using the Uid or something else.