Got session persistence to work (needs improvement)
I got session persistence to work, but it could use a lot of improvement. For now I'm creating a cookie and storing an IdToken in it and then checking if the IdToken is null in the cookie when the user visits the site again. However, the firebase sdk has a CreateSessionCookieAsync method and VerifySessionCookieAsync method which I'm going to use instead.