Updated session persistence to use Firebase SDK Cookies
Updated method of session persistence. Now whenever a user logs in or creates an account, a new session cookie is generated. Each cookie expires in 14 days. When the user logs in again, the firebase SDK verifies if the cookie is still valid and if so, logs the user in. If not, it catches the error and deletes the cookie.