Fixed issue where view would render before cookies were verified
Fixed issue where view would render before cookies were rendered and user data was stored in the session. In the home controller, I made the StoreUserDataInSession an Asynchronous task that the Action Results had to await before rendering the views. This should ensure that the cookies are always verified before the view renders and the login/sign up buttons don't appear if the user is logged in.