Started setting up DbContext and Entities for database
Created a UserAnime class that will store some anime information that will be displayed on the Vault page. It contains the Anime Id from the myanimelist api so I can retrieve the rest of the data using that. I also started setting up the relationships between the entities in the DbContext.