Created an AddToList button and action for anime details page.
I created an add to list button on the anime details page as well as an action result that it calls when pressed. The action result checks if the user is still authenticated and then will add the anime to the users list and reload the page. This still needs implementation. I also renamed AddUserDetailsToSession() to ValidateUserSession() and got rid of the session variables check so now it only checks if the session variable is still valid. I felt this was more secure.