Tested Adding To User List and Fixed UserDataService NavigationProperty Issue
Tested adding a few UserAnimes to the database which seems to work as intended. There was a slight issue where the posters on the Vault view wouldn't show because the User navigation properties weren't being loaded properly, but I fixed this issue by changing the logic in the UserDataService to use both Include and ThenInclude statements so that the navigation properties of the navigation properties were loaded.