User can now change rating and watch status of anime in their list via AJAX
Added backend logic to the select properties for rating and watch status on the anime details page. When the user changes one of these properties, the server sends an ajax request and calls my ActionResults in the HomeController. The action results attempt to update these values in the database and returns a success status.