Created input for episodes watched on anime details page
Created a number input for episodes watched on the anime details page. It uses a debounce timer and updates the database 500ms after the user changes it's input. It also checks whether the input the user entered was valid or note. I also styled the input to match the theme of the site.