Created remove from vault button on anime details page
Created a remove from vault button that removes an anime from a users list and the UserAnime entry from the database. I also created a unit test to test if this works and all tests passed. I still need to fix the issue where if a show is still airing and the total episodes is unknown how to update the watched episdoes, but other than that the site should be ready to publish.