Fixed issue where API wouldn't add animes to user list
Fixed the issue where the AddAnimeToList endpoint was returning an internal server error. This was because the DbContext was not saving changes after adding the poster and start season to the db. I also deleted all unnecessary nuget packages now that the main website doesn't need Entity Framework anymore, and published both the website and the api to azure.